Code overhaul and memory violation fixes
Lots of changes on the code structure to make the code less bulky and more logical.

Rework of `slog_itoa_pad ()` function, which lead to a reduce in the amount of unnecessary dynamic allocations in `slog_fmt_get_str ()` function and fixed an out-of-bounds array write.
Addition of `slog_puts ()` function, which acts as a replacement for puts () from stdio.
Rename of `slog.h` definitions to make them more compatible with other compilers.
Fixes of some grammatical errors in files.